Move .posttag to .link.posttag
