+++ src/main/java/com/kaba4cow/imgxiv/domain/link/posttag/PostTagRepository.java
@@ -1,4 +1,4 @@
-package com.kaba4cow.imgxiv.domain.posttag;
+package com.kaba4cow.imgxiv.domain.link.posttag;
 
 import org.springframework.data.jpa.repository.JpaRepository;
