+++ src/main/java/com/kaba4cow/imgxiv/domain/post/Post.java
@@ -9,7 +9,7 @@ import org.hibernate.annotations.OnDeleteAction;
 
 import com.kaba4cow.imgxiv.domain.base.EntityWithId;
 import com.kaba4cow.imgxiv.domain.embeddable.CreatedAt;
-import com.kaba4cow.imgxiv.domain.posttag.PostTag;
+import com.kaba4cow.imgxiv.domain.link.posttag.PostTag;
 import com.kaba4cow.imgxiv.domain.tag.Tag;
 import com.kaba4cow.imgxiv.domain.user.User;
